api/requirements.txt
click==8.1.3
cryptography==3.4.7
Flask==1.1.2
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow-sqlalchemy==0.28.1
PyJWT==1.7.1
PyMySQL==1.0.2
python-dotenv==0.20.0
pip==22.0.4
setuptools==50.3.1
Werkzeug==1.0.1